requirements.txt
pytube
moviepy
openai-whisper
transformers
torch
ffmpeg-python
numpy
yt-dlp